refactor: 리스너 콜백을 콜백 메서드로 리팩터링
