+++ backend/src/main/java/com/staccato/moment/domain/Moment.java
@@ -6,7 +6,6 @@ import java.time.temporal.ChronoUnit;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
@@ -15,10 +14,12 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.PrePersist;
+import jakarta.persistence.PreRemove;
+import jakarta.persistence.PreUpdate;
 import com.staccato.config.domain.BaseEntity;
 import com.staccato.exception.StaccatoException;
 import com.staccato.memory.domain.Memory;
-import com.staccato.moment.domain.support.MomentEntityListener;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -27,7 +28,6 @@ import lombok.NonNull;
 
 @Entity
 @Getter
-@EntityListeners({MomentEntityListener.class})
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Moment extends BaseEntity {
     @Id
@@ -97,5 +97,12 @@ public class Moment extends BaseEntity {
 
     public void changeFeeling(Feeling feeling) {
         this.feeling = feeling;
+    }
+
+    @PrePersist
+    @PreUpdate
+    @PreRemove
+    public void touchForWrite() {
+        memory.setUpdatedAt(LocalDateTime.now());
     }
 }
