+++ backend/src/test/java/com/staccato/moment/domain/MomentTest.java
@@ -4,13 +4,25 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.transaction.annotation.Transactional;
 import com.staccato.exception.StaccatoException;
+import com.staccato.fixture.Member.MemberFixture;
 import com.staccato.fixture.memory.MemoryFixture;
+import com.staccato.fixture.moment.MomentFixture;
+import com.staccato.member.domain.Member;
+import com.staccato.member.repository.MemberRepository;
 import com.staccato.memory.domain.Memory;
+import com.staccato.memory.repository.MemoryRepository;
+import com.staccato.moment.repository.MomentRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
@@ -105,5 +117,75 @@ class MomentTest {
                 .build())
                 .isInstanceOf(StaccatoException.class)
                 .hasMessageContaining("추억에 포함되지 않는 날짜입니다.");
+    }
+
+    @Nested
+    @Transactional
+    @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
+    class TouchForWriteTest {
+        @Autowired
+        private MemberRepository memberRepository;
+        @Autowired
+        private MemoryRepository memoryRepository;
+        @Autowired
+        private MomentRepository momentRepository;
+        @PersistenceContext
+        private EntityManager entityManager;
+
+        @DisplayName("Moment 생성 시 Memory의 updatedAt이 갱신된다.")
+        @Test
+        void updateMemoryUpdatedDateWhenMomentCreated() {
+            // given
+            Member member = memberRepository.save(MemberFixture.create());
+            Memory memory = memoryRepository.save(MemoryFixture.createWithMember(member));
+            LocalDateTime beforeCreate = memory.getUpdatedAt();
+
+            // when
+            momentRepository.save(MomentFixture.create(memory));
+            entityManager.flush();
+            entityManager.refresh(memory);
+            LocalDateTime afterCreate = memory.getUpdatedAt();
+
+            // then
+            assertThat(afterCreate).isAfter(beforeCreate);
+        }
+
+        @DisplayName("Moment 수정 시 Memory의 updatedAt이 갱신된다.")
+        @Test
+        void updateMemoryUpdatedDateWhenMomentUpdated() {
+            // given
+            Member member = memberRepository.save(MemberFixture.create());
+            Memory memory = memoryRepository.save(MemoryFixture.createWithMember(member));
+            Moment moment = momentRepository.save(MomentFixture.create(memory));
+            LocalDateTime beforeUpdate = memory.getUpdatedAt();
+
+            // when
+            moment.changeFeeling(Feeling.ANGRY);
+            entityManager.flush();
+            entityManager.refresh(memory);
+            LocalDateTime afterUpdate = memory.getUpdatedAt();
+
+            // then
+            assertThat(afterUpdate).isAfter(beforeUpdate);
+        }
+
+        @DisplayName("Moment 삭제 시 Memory의 updatedAt이 갱신된다.")
+        @Test
+        void updateMemoryUpdatedDateWhenMomentDeleted() {
+            // given
+            Member member = memberRepository.save(MemberFixture.create());
+            Memory memory = memoryRepository.save(MemoryFixture.createWithMember(member));
+            Moment moment = momentRepository.save(MomentFixture.create(memory));
+            LocalDateTime beforeDelete = memory.getUpdatedAt();
+
+            // when
+            momentRepository.delete(moment);
+            entityManager.flush();
+            entityManager.refresh(memory);
+            LocalDateTime afterDelete = memory.getUpdatedAt();
+
+            // then
+            assertThat(afterDelete).isAfter(beforeDelete);
+        }
     }
 }
